Update init.pp
This is a simple enhancement came about when we needed to centrally manage repos on our CentOS servers. The yum class we used previously would purge any repos not managed by puppet, but the current implementation does not. The addition of the one variable and file resource allows for puppet to purge those repos not managed by puppet.  It defaults to false, but when set to true removes all repos not managed by the yum class.